+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Utility/HDUtils.cs
@@ -24,6 +24,8 @@ namespace UnityEngine.Rendering.HighDefinition
         static internal HDAdditionalLightData s_DefaultHDAdditionalLightData { get { return ComponentSingleton<HDAdditionalLightData>.instance; } }
         /// <summary>Default HDAdditionalCameraData</summary>
         static internal HDAdditionalCameraData s_DefaultHDAdditionalCameraData { get { return ComponentSingleton<HDAdditionalCameraData>.instance; } }
+        
+        static List<CustomPassVolume> m_TempCustomPassVolumeList = new List<CustomPassVolume>();
 
         static Texture3D m_ClearTexture3D;
         static RTHandle m_ClearTexture3DRTH;
@@ -727,6 +729,24 @@ namespace UnityEngine.Rendering.HighDefinition
             float distanceToCamera = Vector3.Magnitude(position1 - position2);
             float distanceFade = ComputeLinearDistanceFade(distanceToCamera, fadeDistance);
             return distanceFade * weight;
+        }
+
+        internal static bool WillCustomPassBeExecuted(HDCamera hdCamera, CustomPassInjectionPoint injectionPoint)
+        {
+            if (!hdCamera.frameSettings.IsEnabled(FrameSettingsField.CustomPass))
+                return false;
+
+            bool executed = false;
+            CustomPassVolume.GetActivePassVolumes(injectionPoint, m_TempCustomPassVolumeList);
+            foreach(var customPassVolume in m_TempCustomPassVolumeList)
+            {
+                if (customPassVolume == null)
+                    return false;
+
+                executed |= customPassVolume.WillExecuteInjectionPoint(hdCamera);
+            }
+
+            return executed;
         }
 
         internal static bool PostProcessIsFinalPass()